test(e2e): share the private-DB fixtures through conftest
